render about page at / for unauthed
